+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.3] - 2022-05-30
 ### Added
 - More complex neighborhood data augmentation
 - Rule return type
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fallback to `default_tokenizer()` for `sklearn.CountVectorizer` and `sklearn.TfidfVectorizer`
+- Bugfixes in feature selection when `n_features` >= `n_samples`
 
 ## [0.6.2] - 2022-04-06
 ### Changed
@@ -271,6 +274,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [Unreleased]: https://git.science.uu.nl/m.j.robeer/text_explainability
+[0.6.3]: https://pypi.org/project/text-explainability/0.6.3/
 [0.6.2]: https://pypi.org/project/text-explainability/0.6.2/
 [0.6.1]: https://pypi.org/project/text-explainability/0.6.1/
 [0.6.0]: https://pypi.org/project/text-explainability/0.6.0/